+upd room deleting +new features

+++ Hotel/src/main/java/com/ua/rd/Hotel/rest/RoomController.java
@@ -43,8 +43,13 @@ public class RoomController {
     }
 
     @DeleteMapping("/room/{id}")
-    public void deleteRoom(@PathVariable(value = "id") Long id) {
+    public ResponseEntity<Object> deleteRoom(@PathVariable(value = "id") Long id) {
+        Optional<Room> room = roomService.findByIdUpdate(id);
+        if (room.isEmpty()) {
+            return ResponseEntity.notFound().build();
+        }
         roomService.deleteById(id);
+        return null;
     }
 
 
@@ -66,6 +71,29 @@ public class RoomController {
              @RequestParam("checkOutDate")
              LocalDate checkOut) {
         return roomService.findRoomsNotReservationListInRange(checkIn, checkOut);
+    }
+
+    @PutMapping("/room/{id}")
+    public ResponseEntity<Void> updateRoom(@PathVariable("id") Long id, @RequestBody RoomDto roomDto) {
+        Optional<Room> optionalRoom = roomService.findByIdUpdate(id);
+
+        if (optionalRoom.isEmpty()) {
+            return ResponseEntity.notFound().build();
+        }
+
+        Room room = optionalRoom.get();
+
+
+        room.setName(roomDto.getName());
+        room.setCapacity(roomDto.getCapacity());
+        room.setFloor(roomDto.getFloor());
+        room.setNumberOfBeds(roomDto.getNumberOfBeds());
+
+
+
+        roomService.save(room);
+
+        return ResponseEntity.ok().build();
     }
 
 }
